Refactor backend into packages
