+++ backend/services/db.py
@@ -1,0 +1,29 @@
+import os
+import json
+from pathlib import Path
+
+from dotenv import load_dotenv
+from motor.motor_asyncio import AsyncIOMotorClient
+import firebase_admin
+from firebase_admin import credentials
+
+ROOT_DIR = Path(__file__).resolve().parent.parent
+load_dotenv(ROOT_DIR / '.env')
+
+mongo_url = os.environ['MONGO_URL']
+client = AsyncIOMotorClient(mongo_url)
+db = client[os.environ['DB_NAME']]
+
+
+def init_firebase() -> None:
+    """Initialize Firebase Admin SDK if credentials are provided."""
+    try:
+        firebase_service_account = json.loads(os.environ.get('FIREBASE_SERVICE_ACCOUNT', '{}'))
+        if firebase_service_account:
+            cred = credentials.Certificate(firebase_service_account)
+            firebase_admin.initialize_app(cred)
+            print("Firebase Admin SDK initialized successfully")
+        else:
+            print("Firebase credentials not found - using placeholder mode")
+    except Exception as e:
+        print(f"Firebase initialization failed: {e}")
